+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -79,5 +79,26 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-ManagedIdentity");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestOsSku()
+        {
+            TestRunner.RunTestScript("Test-OSSku");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNodeLabelsAndTags()
+        {
+            TestRunner.RunTestScript("Test-NodeLabels-Tags");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNodeTaints()
+        {
+            TestRunner.RunTestScript("Test-NodeTaints");
+        }
     }
 }