+++ src/ManagedServiceIdentity/exports/Get-AzFederatedIdentityCredentials.ps1
@@ -45,7 +45,7 @@ INPUTOBJECT <IManagedServiceIdentity>: Identity Parameter
   [Scope <String>]: The resource provider scope of the resource. Parent resource being extended by Managed Identities.
   [SubscriptionId <String>]: The Id of the Subscription to which the identity belongs.
 .Link
-https://docs.microsoft.com/powershell/module/az.managedserviceidentity/get-azfederatedidentitycredentials
+https://learn.microsoft.com/powershell/module/az.managedserviceidentity/get-azfederatedidentitycredentials
 #>
 function Get-AzFederatedIdentityCredentials {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.ManagedServiceIdentity.Models.Api20220131Preview.IFederatedIdentityCredential])]