+++ src/Resources/Authorization.Autorest/exports/Get-AzRoleAssignmentScheduleRequest.ps1
@@ -53,7 +53,7 @@ INPUTOBJECT <IAuthorizationIdentity>: Identity Parameter
   [RoleManagementPolicyName <String>]: The name (guid) of the role management policy to get.
   [Scope <String>]: The scope of the role management policy.
 .Link
-https://docs.microsoft.com/powershell/module/az.resources/get-azroleassignmentschedulerequest
+https://learn.microsoft.com/powershell/module/az.resources/get-azroleassignmentschedulerequest
 #>
 function Get-AzRoleAssignmentScheduleRequest {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.Resources.Authorization.Models.Api20201001Preview.IRoleAssignmentScheduleRequest])]