+++ src/ServiceBus/ServiceBus.Autorest/test/Remove-AzServiceBusPrivateEndpointConnection.Tests.ps1
@@ -16,12 +16,12 @@ if(($null -eq $TestName) -or ($TestName -contains 'Remove-AzServiceBusPrivateEnd
 
 Describe 'Remove-AzServiceBusPrivateEndpointConnection' {
     $listOfPrivateEndpoints = Get-AzServiceBusPrivateEndpointConnection -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace
-    It 'Delete' {
+    It 'Delete'  {
         Remove-AzServiceBusPrivateEndpointConnection -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace -Name $listOfPrivateEndpoints[0].Name
         { Get-AzServiceBusPrivateEndpointConnection -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace -Name $listOfPrivateEndpoints[0].Name } | Should -Throw
     }
 
-    It 'DeleteViaIdentity' {
+    It 'DeleteViaIdentity'  {
         $privateEndpoint = Get-AzServiceBusPrivateEndpointConnection -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace -Name $listOfPrivateEndpoints[1].Name
         Remove-AzServiceBusPrivateEndpointConnection -InputObject $privateEndpoint
         { Get-AzServiceBusPrivateEndpointConnection -InputObject $privateEndpoint } | Should -Throw