+++ src/StackHCI/exports/Remove-AzStackHciExtension.ps1
@@ -41,7 +41,7 @@ INPUTOBJECT <IStackHciIdentity>: Identity Parameter
   [ResourceGroupName <String>]: The name of the resource group. The name is case insensitive.
   [SubscriptionId <String>]: The ID of the target subscription.
 .Link
-https://docs.microsoft.com/powershell/module/az.stackhci/remove-azstackhciextension
+https://learn.microsoft.com/powershell/module/az.stackhci/remove-azstackhciextension
 #>
 function Remove-AzStackHciExtension {
 [OutputType([System.Boolean])]