+++ src/Aks/Aks/ChangeLog.md
@@ -19,6 +19,19 @@
 -->
 ## Upcoming Release
 
+## Version 5.1.0
+* Bumped API version to 2022-09-01
+* Added parameter `-NodeOsSKU` for `New-AzAksCluster` and parameter `-OsSKU` for `New-AzAksNodePool`
+* Added parameter `-Mode` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added property `-NodeImageVersion` for the output of `Get-AzAksNodePool`[#19893]
+* Added parameter `-NodePoolLabel` for `Set-AzAksCluster`, `-NodeLabel` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-NodePoolTag` for `New-AzAksCluster` and `Set-AzAksCluster`, `-Tag` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-NodeTaint` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-AksCustomHeader` for `Set-AzAksCluster`, `New-AzAksNodePool` and `Update-AzAksNodePool`
+
+## Version 5.0.1
+* Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]
+
 ## Version 5.0.0
 * [Breaking Change] Removed the alias `Install-AzAksKubectl` of `Install-AzAksCliTool`.